requirements.txt
Flask
pymysql
python-dotenv
cryptography
matplotlib